backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mediapipe>=0.10.30
opencv-python-headless==4.9.0.80
numpy<2
scikit-learn>=1.4.0
hypothesis==6.100.0
pytest==8.2.0
websockets==12.0
python-multipart==0.0.9
openai>=1.0.0
python-dotenv>=1.0.0
